Palo Alto Networks/requirements.txt
netmiko==4.5.0
pan-os-python==1.12.3
pan-python==0.17.0
requests==2.32.4
setuptools>=75.1.0
